AwesomeEquations

Documentation for AwesomeEquations.

AwesomeEquations.Dual

AwesomeEquations.finite_diff

AwesomeEquations.forward_diff

AwesomeEquations.forward_euler

AwesomeEquations.rk4

AwesomeEquations.Dual — Type

Dual(a, b)

Dual number with value a and derivative b.

A dual number $x = a + b \epsilon$ is defined such that $\epsilon^2 = 0$ and $f(a + b \epsilon) = f(a) + f'(a) b \epsilon$ for all $f$. The derivative of a function is thus given by $f'(a) = y$ when $f(a + \epsilon) = x + y \epsilon$.

AwesomeEquations.finite_diff — Method

finite_diff(f, x; h = sqrt(eps(typeof(x))))

Compute the derivative of f at x using a central finite difference (second order).

AwesomeEquations.forward_diff — Method

forward_diff(f, x)

Compute the derivative of f at x using forward mode automatic differention (with dual numbers).

AwesomeEquations.forward_euler — Function

forward_euler(f, p, u₀, T, Δt = 0.1)

Solve ODE $du/dt = f(u, p, t)$ with forward Euler, where $p$ are parameters.

AwesomeEquations.rk4 — Function

rk4(f, p, u₀, T, Δt = 0.1)

Solve ODE $du/dt = f(u, p, t)$ with RK4, where $p$ are parameters.